[fix] changed integer type for columns

+++ src/database/migrations/1638217974847-UserTable.ts
@@ -8,7 +8,7 @@ export class UserTable1638217974847 implements MigrationInterface {
         columns: [
           {
             name: 'id',
-            type: 'int4',
+            type: 'bigint',
             isPrimary: true,
             isGenerated: true,
             generationStrategy: 'increment',
